refactor: make entity dir
