+++ server/src/cat/cat.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CatService } from './cat.service';
 import { CatController } from './cat.controller';
-import { Cat } from './cat.entity';
+import { Cat } from './entity/cat.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Cat])],
