+++ server/src/cat/cat.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CatService } from './cat.service';
-import { Cat } from './cat.entity';
+import { Cat } from './entity/cat.entity';
 
 describe('CatService', () => {
   let service: CatService;
